check_openshift: Modification of check pod to check replicas
Check pod has been modified to check directly all the replicas
configuration in the default namespace instead of the deployment
configuration. Indeed all information: spec & status of the replicas are
available in the replicationcontrollers, thus giving access to the
needed information directly.

The namespace is a parameter, so we can imagine to extend this check for
other namespace and not only the default one.

Should close https://github.com/redhat-cip/rcip-openshift-ansible/issues/14